pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "audio-transcribe"
version = "0.1.0"
description = "Unified audio transcription tool supporting multiple APIs"
authors = [
    {name = "Audio Transcribe Team"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "click",
    "loguru",
    "pydub",
    "python-dotenv",
    "requests",
    "assemblyai",
    "groq",
    "openai>=1.78.0",
    "questionary>=2.1.0",
    "rich>=14.2.0",
]
readme = "README.md"

[project.optional-dependencies]
dev = [
    "pyinstaller",
]

[project.scripts]
transcribe = "audio_transcribe.cli:main"

[tool.setuptools]
packages = ["audio_transcribe", "audio_transcribe.utils", "audio_transcribe.transcribe_helpers"]

[tool.pyinstaller]
hiddenimports = [
    "loguru",
    "assemblyai",
    "groq",
    "python-dotenv",
    "requests",
    "pydub"
]
